requirements.txt
flask==2.1.2
Werkzeug==2.1.2
flask-restx==0.5.1
paho-mqtt==1.6.0
requests==2.27.1
pymongo==4.1.1
dnspython==2.2.1
mysqlclient==2.0.3
python-telegram-bot==13.11
gunicorn==20.1.0
